+++ common-custom-user-data-gradle-plugin/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 - Add custom value and search link for GitHub Actions run ID.
 - Fix generation of search links for custom values (#187).
+- Redact any `user:password@` portion of the 'Git repository' custom value.
 
 ## [1.4.2] - 2021-07-13
 - Fix configuration cache compatibility when capturing test maxParallelForks.